lambda/requirements.txt
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.0
python-multipart==0.0.6
transformers==4.35.0
pillow==10.1.0
numpy==1.24.3
boto3==1.29.0
onnxruntime==1.16.3
